+++ ipc/provider/Cargo.toml
@@ -41,7 +41,8 @@ fvm_ipld_encoding = { workspace = true }
 primitives = { workspace = true }
 
 ipc-identity = { path = "../identity", features = ["with-ethers"] }
-ipc-sdk = { path = "../sdk" }
+ipc-sdk = { workspace = true }
+ipc_actors_abis = { workspace = true }
 
 [dev-dependencies]
 tempfile = { workspace = true }
