--- ipc/provider/src/manager/evm/convert/checkpoint.rs
@@ -1,2 +1,0 @@
-// Copyright 2022-2023 Protocol Labs
-// SPDX-License-Identifier: MIT
